feat: add experimental module - chore: move triplet related handlers to experimental - chore: general linter pass
